Change directory structure and implement IIFE pattern
