+++ src/index.ts
@@ -2,6 +2,6 @@ export { RoarFirekit } from './firekit';
 export { RoarAppkit } from './firestore/app/appkit';
 export { RoarAppUser } from './firestore/app/user';
 export { RoarTaskVariant } from './firestore/app/task';
-export { canUser } from './permissions/permissionsManager';
+export { PermissionsService } from './service/permissions-service';
 export { emptyOrg, emptyOrgList, getTreeTableOrgs, initializeFirebaseProject } from './firestore/util';
 export * from './firestore/query-assessment';
