+++ src/service/permissions-service.test.ts
@@ -1,12 +1,12 @@
-import { canUser } from '../permissions/permissionsManager';
+import { PermissionsService } from './permissions-service';
 
-const ADMIN_TOKEN =
+const MOCK_ADMIN_TOKEN =
   'eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJSUE1fVGVzdF9Ub2tlbiIsImlhdCI6MTczOTU1ODk1MiwiZXhwIjoxNzcxMDk3ODkyLCJhdWQiOiJyb2FyLmVkdWNhdGlvbiIsInN1YiI6InRlc3RUb2tlbjEiLCJyb2xlIjoiQURNSU4ifQ.Q32pjnv5RVljsQYMxU7d40N3DOGvp7plkWOLQwlcdJQ';
-const STUDENT_TOKEN =
+const MOCK_STUDENT_TOKEN =
   'eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJSUE1fVGVzdF9Ub2tlbiIsImlhdCI6MTczOTU1ODk1MiwiZXhwIjoxNzcxMDk3ODkyLCJhdWQiOiJyb2FyLmVkdWNhdGlvbiIsInN1YiI6InRlc3RUb2tlbjIiLCJyb2xlIjoiU1RVREVOVCJ9.yp4zLv4_nVujW8XPfTvoQjfpmAEVcGhyZL0Hm_ubF6k';
-const PLATFORM_ADMIN_TOKEN =
+const MOCK_PLATFORM_ADMIN_TOKEN =
   'eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJSUE1fVGVzdF9Ub2tlbiIsImlhdCI6MTczOTU1ODk1MiwiZXhwIjoxNzcxMDk3ODkyLCJhdWQiOiJyb2FyLmVkdWNhdGlvbiIsInN1YiI6InRlc3RUb2tlbjQiLCJyb2xlIjoiUExBVEZPUk1fQURNSU4ifQ.vpW-SUhdScLFUlA5-wP8ZYLd8pDFMGvJQqyef4R7BP0';
-const SUPER_ADMIN_TOKEN =
+const MOCK_SUPER_ADMIN_TOKEN =
   'eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJSUE1fVGVzdF9Ub2tlbiIsImlhdCI6MTczOTU1ODk1MiwiZXhwIjoxNzcxMDk3ODkyLCJhdWQiOiJyb2FyLmVkdWNhdGlvbiIsInN1YiI6InRlc3RUb2tlbjMiLCJyb2xlIjoiU1VQRVJfQURNSU4ifQ.Q6AEfvL2pKi9l4tiSCwwBhL5HgjG8__FTDKbZR-usu4';
 
 describe('canUser', () => {
@@ -21,7 +21,7 @@ describe('canUser', () => {
     ];
 
     for (const action of permissions) {
-      const canTakeAction = canUser(STUDENT_TOKEN, action.action);
+      const canTakeAction = PermissionsService.canUser(MOCK_STUDENT_TOKEN, action.action);
       expect(canTakeAction).toBe(action.expected);
     }
   });
@@ -36,7 +36,7 @@ describe('canUser', () => {
     ];
 
     for (const action of permissions) {
-      const canTakeAction = canUser(ADMIN_TOKEN, action.action);
+      const canTakeAction = PermissionsService.canUser(MOCK_ADMIN_TOKEN, action.action);
       expect(canTakeAction).toBe(action.expected);
     }
   });
@@ -51,7 +51,7 @@ describe('canUser', () => {
     ];
 
     for (const action of permissions) {
-      const canTakeAction = canUser(PLATFORM_ADMIN_TOKEN, action.action);
+      const canTakeAction = PermissionsService.canUser(MOCK_PLATFORM_ADMIN_TOKEN, action.action);
       expect(canTakeAction).toBe(action.expected);
     }
   });
@@ -66,7 +66,7 @@ describe('canUser', () => {
     ];
 
     for (const action of permissions) {
-      const canTakeAction = canUser(SUPER_ADMIN_TOKEN, action.action);
+      const canTakeAction = PermissionsService.canUser(MOCK_SUPER_ADMIN_TOKEN, action.action);
       expect(canTakeAction).toBe(action.expected);
     }
   });
